==> 2. HandlingTestData.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns

test = pd.read_csv('test.csv')

test.shape        # Does not have SalePrice

test[test.columns[test.isnull().any()]].isnull().sum().sort_values(ascending=False)

test.drop(['PoolQC','MiscFeature','Alley','Fence'], axis=1, inplace=True)

test.shape

test['LotFrontage'] = test['LotFrontage'].fillna(test['LotFrontage'].mean())

test['FireplaceQu'] = test['FireplaceQu'].fillna(test['FireplaceQu'].mode()[0])

GarageData = ['GarageType', 'GarageFinish', 'GarageQual', 'GarageCond']
test[GarageData] = test[GarageData].fillna('None')
test['GarageYrBlt'] = test['GarageType'].fillna(0)

BasementData = ['BsmtExposure', 'BsmtFinType2', 'BsmtFinType1', 'BsmtCond', 'BsmtQual']
test[BasementData] = test[BasementData].fillna('None')

test['MasVnrArea'] = test['MasVnrArea'].fillna(test['MasVnrArea'].mean())

test['MasVnrType'] = test['MasVnrType'].fillna(test['MasVnrType'].mode()[0])

test['Electrical'] = test['Electrical'].fillna(test['Electrical'].mode()[0])

test[test.columns[test.isnull().any()]].isnull().sum().sort_values(ascending=False)

def fillmode(cols):
    df = test
    for col in cols:
        df[col]=df[col].fillna(df[col].mode()[0])

obj_cols=['MSZoning', 'Utilities', 'Functional', 'Exterior1st', 'Exterior2nd', 'KitchenQual', 'SaleType']
fillmode(obj_cols)

def fillmean(cols):
    df = test
    for col in cols:
        df[col]=df[col].fillna(df[col].mean())

int_cols=['BsmtFullBath', 'BsmtHalfBath', 'BsmtFinSF1', 'BsmtFinSF2', 'BsmtUnfSF', 'TotalBsmtSF', 'GarageCars', 'GarageArea']
fillmean(int_cols)

test[test.columns[test.isnull().any()]].isnull().sum().sort_values(ascending=False) # No more null columns

test.drop('Id', axis=1, inplace=True)

# After this we can concatenate both train and test sets and then go for one hot encoding

test.to_csv('test_no_missing.csv', index=False)

test.shape
