==> 3. GetDummies.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns

train_df = pd.read_csv('train_no_missing.csv')

test_df = pd.read_csv('test_no_missing.csv')

train_df.shape

test_df.shape

train_df = train_df.drop(['Id', 'SalePrice'], axis=1)

train_df.shape

final_df = pd.concat([train_df,test_df], axis=0)

final_df.shape

final_df = pd.get_dummies(final_df, drop_first=True)

final_df.shape

df_train = final_df.iloc[:1460, :]
df_test = final_df.iloc[1460:, :]

df_train.shape

df_test.shape

# We are going to deal with only df_train
# df_test does not have a SalePrice column so we will have to submit it to Kaggle to get the results

df_train.to_csv('train_dummies.csv', index=False)
df_test.to_csv('test_dummies.csv', index=False)
