==> 4. LinearRegression.ipynb <==
# Linear Regression

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns

X = pd.read_csv('train_dummies.csv')

y = pd.read_csv('train.csv')['SalePrice']

from sklearn.model_selection import train_test_split
X_train,X_test,y_train,y_test = train_test_split(X, y, test_size=0.3, random_state=101)

from sklearn.linear_model import LinearRegression
linear_model = LinearRegression()

linear_model.fit(X_train,y_train)

lr_preds = linear_model.predict(X_test)

from sklearn.metrics import mean_absolute_error, mean_squared_error

mae = mean_absolute_error(y_test, lr_preds)
print(mae)

rmse = np.sqrt(mean_squared_error(y_test, lr_preds))
print(rmse)

# Let's import main test date
final_test = pd.read_csv('test_dummies.csv')

final_lr = linear_model.predict(final_test)

final_lr

submission = pd.concat([pd.read_csv('test.csv')['Id'],pd.DataFrame(final_lr)], axis=1)
submission.columns = ['Id', 'SalePrice']

submission.to_csv('LRsubmission.csv', index=False)
