==> 5. RidgeRegression.ipynb <==
# Polynomial Regression
# We cannot go for polynomial regression directly on the categorical features, we need to convert them to integers first
# But if we go for polynomial features on get dummies, the dimensionality of the dataset would increase widely
# We must go for feature selection - Using main dataset

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns

df = pd.read_csv('train_no_missing.csv')

df.shape

df.corr()['SalePrice'].sort_values(ascending=False)

# Let's use only those features which have more than 30% correlation with SalePrice
# Notice that all are integer features

feature_df = df[['OverallQual', 'GrLivArea', 'GarageCars', 'GarageArea', 'TotalBsmtSF', '1stFlrSF', 'FullBath', 'TotRmsAbvGrd',
                'YearBuilt', 'YearRemodAdd', 'MasVnrArea', 'Fireplaces', 'BsmtFinSF1', 'LotFrontage', 'WoodDeckSF', 
                '2ndFlrSF', 'OpenPorchSF']]

from sklearn.preprocessing import PolynomialFeatures

poly_converter = PolynomialFeatures(degree=3, include_bias=False)

polynomial_features = poly_converter.fit_transform(feature_df)

polynomial_features.shape

from sklearn.model_selection import train_test_split

X_train, X_test, y_train, y_test = train_test_split(polynomial_features, df['SalePrice'], test_size=0.3, random_state=101)

X_train.shape

from sklearn.preprocessing import StandardScaler

scaler = StandardScaler()

scaler.fit(X_train)

X_train = scaler.transform(X_train)
X_test = scaler.transform(X_test)

from sklearn.linear_model import Ridge,RidgeCV

ridge_cv_model = RidgeCV(alphas=(0.1, 1.0, 10.0))

ridge_cv_model.fit(X_train, y_train)

ridge_predictions = ridge_cv_model.predict(X_test)

from sklearn.metrics import mean_absolute_error, mean_squared_error

mae = mean_absolute_error(y_test, ridge_predictions)
mae

rmse = np.sqrt(mean_squared_error(y_test, ridge_predictions))
rmse

ridge_cv_model.alpha_

# Since we haven't used a large number of features, the performance hasn't improved much compared to LR

df_test = pd.read_csv('test_no_missing.csv')[['OverallQual', 'GrLivArea', 'GarageCars', 'GarageArea', 'TotalBsmtSF', '1stFlrSF', 'FullBath', 'TotRmsAbvGrd', 'YearBuilt', 'YearRemodAdd', 'MasVnrArea', 'Fireplaces', 'BsmtFinSF1', 'LotFrontage', 'WoodDeckSF', '2ndFlrSF', 'OpenPorchSF']]

test_features = poly_converter.transform(df_test)

scaled_test = scaler.transform(test_features)

final_predictions = ridge_cv_model.predict(scaled_test)

final_predictions

final_ridge = pd.concat([pd.read_csv('test.csv')['Id'],pd.DataFrame(final_predictions)], axis=1)

final_ridge.columns = ['Id', 'SalePrice']

final_ridge.to_csv('RidgeSubmission.csv')

# We can also try ElasticNet

from sklearn.linear_model import ElasticNetCV

elastic_model = ElasticNetCV(l1_ratio=[.1, .5, .7, .9, .95, .99, 1], eps=0.001, n_alphas=1000, max_iter=1000000)

elastic_model.fit(X_train, y_train)

elastic_model.alpha_

elastic_predictions = elastic_model.predict(X_test)

mae = mean_absolute_error(y_test, elastic_predictions)
mae

rmse = np.sqrt(mean_squared_error(y_test, elastic_predictions))
rmse

# Elastic performed worse than Ridge
