==> 8. Tree Based Regressor.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns

X = pd.read_csv('train_dummies.csv')
y = pd.read_csv('train.csv')['SalePrice']

from sklearn.model_selection import train_test_split
X_train,X_test,y_train,y_test = train_test_split(X, y, test_size=0.3, random_state=101)

from sklearn.preprocessing import StandardScaler
scaler = StandardScaler()
scaler.fit(X_train)
X_train = scaler.transform(X_train)
X_test = scaler.transform(X_test)

from sklearn.tree import DecisionTreeRegressor

model = DecisionTreeRegressor(random_state=101)

model.fit(X_train, y_train)

dc_preds = model.predict(X_test)

from sklearn.metrics import mean_absolute_error, mean_squared_error

mae = mean_absolute_error(y_test, dc_preds)
mae

rmse = np.sqrt(mean_squared_error(y_test, dc_preds))
rmse

final_test = pd.read_csv('test_dummies.csv')
final_test = scaler.transform(final_test)

final_preds = model.predict(final_test)

final_preds

submission = pd.concat([pd.read_csv('test.csv')['Id'],pd.DataFrame(final_preds)], axis=1)
submission.columns = ['Id', 'SalePrice']

submission.to_csv('DTsubmission.csv', index=False)

from sklearn.ensemble import RandomForestRegressor

rf_model = RandomForestRegressor(random_state=101)

rf_model.fit(X_train, y_train)

rf_preds = rf_model.predict(X_test)

mae = mean_absolute_error(y_test, rf_preds)
mae

rmse = np.sqrt(mean_squared_error(y_test, rf_preds))
rmse

final_rf_preds = rf_model.predict(final_test)

final_rf_preds

submission = pd.concat([pd.read_csv('test.csv')['Id'],pd.DataFrame(final_rf_preds)], axis=1)
submission.columns = ['Id', 'SalePrice']

submission.to_csv('RFsubmission.csv', index=False)

# Gradient Boost Regressor

from sklearn.ensemble import GradientBoostingRegressor, AdaBoostRegressor
gb_model = GradientBoostingRegressor(random_state=101)

gb_model.fit(X_train, y_train)

gb_preds = gb_model.predict(X_test)

mae = mean_absolute_error(y_test, gb_preds)
mae

rmse = np.sqrt(mean_squared_error(y_test, gb_preds))
rmse

final_gb_preds = gb_model.predict(final_test)

final_gb_preds

submission = pd.concat([pd.read_csv('test.csv')['Id'],pd.DataFrame(final_gb_preds)], axis=1)
submission.columns = ['Id', 'SalePrice']

submission.to_csv('GBsubmission.csv', index=False)

ada_model = AdaBoostRegressor(random_state=101)
ada_model.fit(X_train, y_train)
ada_preds = ada_model.predict(X_test)

mae = mean_absolute_error(y_test, ada_preds)
mae

rmse = np.sqrt(mean_squared_error(y_test, ada_preds))
rmse

final_ada_preds = ada_model.predict(final_test)

submission = pd.concat([pd.read_csv('test.csv')['Id'],pd.DataFrame(final_ada_preds)], axis=1)
submission.columns = ['Id', 'SalePrice']

submission.to_csv('ADAsubmission.csv', index=False)
